game-core/movement fix movement generator part3

+++ game-core/src/main/java/com/github/azeroth/game/movement/spline/MoveSplineInit.java
@@ -139,7 +139,7 @@ public class MoveSplineInit {
 
         unit.sendMessageToSet(packet, true);
 
-        return moveSpline.duration();
+        return (int)moveSpline.duration();
     }
 
     public final void stop() {
